+++ spine-ios/Example/Spine iOS Example/PlayPauseAnimation.swift
@@ -14,13 +14,13 @@ struct PlayPauseAnimation: View {
     
     init() {
         _controller = StateObject(
-            wrappedValue: SpineController { controller in
-                _ = controller.animationState.setAnimationByName(
+            wrappedValue: SpineController(onInitialized: { controller in
+                controller.animationState.setAnimationByName(
                     trackIndex: 0,
                     animationName: "flying",
                     loop: true
                 )
-            }
+            })
         )
     }