+++ spine-ios/CHANGELOG.md
@@ -9,3 +9,4 @@
 - Generate Swift classes from `spine-cpp-lite.h` (#5)
 - Draw `SkeletonData` render commands (#6)
 - Add `BoundsProvider` (#7)
+- Support `SpineController` & `Event` callbacks (#8)